Windows: factor out bridge server+config
